+++ cli/npm/resolvers/global.rs
@@ -26,6 +26,7 @@ use crate::npm::NpmPackageReq;
 use crate::npm::RealNpmRegistryApi;
 
 use super::common::ensure_registry_read_permission;
+use super::common::types_package_name;
 use super::common::InnerNpmPackageResolver;
 
 /// Resolves packages from the global npm cache.
@@ -97,7 +98,7 @@ impl InnerNpmPackageResolver for GlobalNpmPackageResolver {
         }
       }
 
-      let name = format!("@types/{}", name);
+      let name = types_package_name(name);
       let pkg = self
         .resolution
         .resolve_package_from_package(&name, &referrer_pkg_id)?;