+++ cli/tools/standalone.rs
@@ -85,9 +85,9 @@ async fn download_base_binary(
     std::process::exit(1)
   };
 
-  std::fs::create_dir_all(&output_directory)?;
+  std::fs::create_dir_all(output_directory)?;
   let output_path = output_directory.join(binary_path_suffix);
-  std::fs::create_dir_all(&output_path.parent().unwrap())?;
+  std::fs::create_dir_all(output_path.parent().unwrap())?;
   tokio::fs::write(output_path, binary_content).await?;
   Ok(())
 }