+++ ext/flash/01_http.js
@@ -188,8 +188,8 @@
     return str;
   }
 
-  function prepareFastCalls(serverId) {
-    return core.ops.op_flash_make_request(serverId);
+  function prepareFastCalls() {
+    return core.ops.op_flash_make_request();
   }
 
   function hostnameForDisplay(hostname) {
@@ -243,6 +243,7 @@
     i,
     respondFast,
     respondChunked,
+    tryRespondChunked,
   ) {
     // there might've been an HTTP upgrade.
     if (resp === undefined) {
@@ -371,6 +372,9 @@
           }
         } else {
           const reader = respBody.getReader();
+
+          // Best case: sends headers + first chunk in a single go.
+          const { value, done } = await reader.read();
           writeFixedResponse(
             serverId,
             i,
@@ -385,14 +389,23 @@
             false,
             respondFast,
           );
-          while (true) {
-            const { value, done } = await reader.read();
-            await respondChunked(
-              i,
-              value,
-              done,
-            );
-            if (done) break;
+
+          await tryRespondChunked(
+            i,
+            value,
+            done,
+          );
+
+          if (!done) {
+            while (true) {
+              const chunk = await reader.read();
+              await respondChunked(
+                i,
+                chunk.value,
+                chunk.done,
+              );
+              if (chunk.done) break;
+            }
           }
         }
       }
@@ -482,11 +495,15 @@
 
       const serverId = opFn(listenOpts);
       const serverPromise = core.opAsync("op_flash_drive_server", serverId);
-      const listenPromise = PromisePrototypeThen(
-        core.opAsync("op_flash_wait_for_listening", serverId),
-        (port) => {
-          onListen({ hostname: listenOpts.hostname, port });
-        },
+
+      PromisePrototypeCatch(
+        PromisePrototypeThen(
+          core.opAsync("op_flash_wait_for_listening", serverId),
+          (port) => {
+            onListen({ hostname: listenOpts.hostname, port });
+          },
+        ),
+        () => {},
       );
       const finishedPromise = PromisePrototypeCatch(serverPromise, () => {});
 
@@ -502,7 +519,7 @@
             return;
           }
           server.closed = true;
-          core.ops.op_flash_close_server(serverId);
+          await core.opAsync("op_flash_close_server", serverId);
           await server.finished;
         },
         async serve() {
@@ -568,6 +585,7 @@
                           i,
                           respondFast,
                           respondChunked,
+                          tryRespondChunked,
                         ),
                     ),
                     onError,
@@ -585,6 +603,7 @@
                       i,
                       respondFast,
                       respondChunked,
+                      tryRespondChunked,
                     )
                   ).catch(onError);
                   continue;
@@ -603,6 +622,7 @@
                 i,
                 respondFast,
                 respondChunked,
+                tryRespondChunked,
               );
             }
 
@@ -614,10 +634,29 @@
 
       signal?.addEventListener("abort", () => {
         clearInterval(dateInterval);
-        server.close();
+        PromisePrototypeThen(server.close(), () => {}, () => {});
       }, {
         once: true,
       });
+
+      function tryRespondChunked(token, chunk, shutdown) {
+        const nwritten = core.ops.op_try_flash_respond_chuncked(
+          serverId,
+          token,
+          chunk ?? new Uint8Array(),
+          shutdown,
+        );
+        if (nwritten > 0) {
+          return core.opAsync(
+            "op_flash_respond_chuncked",
+            serverId,
+            token,
+            chunk,
+            shutdown,
+            nwritten,
+          );
+        }
+      }
 
       function respondChunked(token, chunk, shutdown) {
         return core.opAsync(
@@ -629,7 +668,7 @@
         );
       }
 
-      const fastOp = prepareFastCalls(serverId);
+      const fastOp = prepareFastCalls();
       let nextRequestSync = () => fastOp.nextRequest();
       let getMethodSync = (token) => fastOp.getMethod(token);
       let respondFast = (token, response, shutdown) =>
@@ -649,8 +688,8 @@
       }
 
       await SafePromiseAll([
-        listenPromise,
         PromisePrototypeCatch(server.serve(), console.error),
+        serverPromise,
       ]);
     };
   }